requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-environ==0.8.1
django-on-heroku==1.1.2
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.0
gunicorn==20.1.0
mango-jwt==1.3.2
passlib==1.7.4
PyJWT==2.3.0
pymongo==3.12.1
pytz==2021.3
whitenoise==5.3.0
